modularizacion de funcion Status
